+++ src/@locale/app/lab/index.js
@@ -4,6 +4,7 @@ import buttons        from './buttons'
 import i18n           from './i18n'
 import mapsMarkers    from './mapsMarkers'
 import mapsModel      from './mapsModel'
+import mapsPicker     from './mapsPicker'
 import mapsView       from './mapsView'
 import metrics        from './metrics'
 import scheduler      from './scheduler'
@@ -21,6 +22,7 @@ const lab =
     i18n           ,
     mapsMarkers    ,
     mapsModel      ,
+    mapsPicker     ,
     mapsView       ,
     metrics        ,
     scheduler      ,